some update

+++ iob_shop/books/routes.py
@@ -2,27 +2,29 @@ from crypt import methods
 from unicodedata import category
 from click import File
 from flask import redirect, request, url_for, render_template, flash, session, current_app
+from sqlalchemy import null
 from iob_shop import db, app, photos
 from iob_shop.admin.routes import publishers
-from .models import Publisher, Category, Author, Book
+from .models import Author_write_book, Book_belongs_to_category, Book_discount, Publisher, Category, Author, Book, Discount
 from .forms import Addbook
 import secrets, os
+from datetime import datetime
 
 @app.route('/')
 @app.route('/home')
 def home():
     page = request.args.get('page', 1, type=int)
-    books = Book.query.filter(Book.stock > 0).paginate(page=page, per_page=2)
-    publishers = Publisher.query.join(Book, (Publisher.id==Book.publisher_id)).all()
-    categories = Category.query.join(Book, (Category.id==Book.category_id)).all()
-    authors = Author.query.join(Book, (Author.id==Book.author_id)).all()
+    books = db.session.query(Book).filter(Book.stock > 0).paginate(page=page, per_page=2)
+    publishers = db.session.query(Publisher).join(Book, (Publisher.id==Book.publisher_id)).all()
+    categories = db.session.query(Category).join(Book, (Category.id==Book.category_id)).all()
+    authors = db.session.query(Author).join(Book, (Author.id==Book.author_id)).all()
     return render_template('books/index.html', books=books, publishers=publishers, categories=categories, authors=authors, title='Home')
 
 @app.route('/publisher/<int:id>')
 def get_publisher(id):
     page = request.args.get('page', 1, type=int)
-    pbypublisher = Book.query.filter_by(publisher_id=id).paginate(page=page, per_page=3)
-    return render_template('books/index.html', books=pbypublisher, title=Publisher.query.get_or_404(id).name)
+    pbypublisher = db.session.query(Book).filter_by(publisher_id=id).paginate(page=page, per_page=3)
+    return render_template('books/index.html', books=pbypublisher, title=db.session.query(Publisher).get_or_404(id).name)
 
 @app.route('/addpublisher', methods=['GET', 'POST'])
 def addpublisher():
@@ -31,22 +33,27 @@ def addpublisher():
         return redirect(url_for('login'))
     if request.method == 'POST':
         getpublisher = request.form.get('publisher')
-        publisher = Publisher(name=getpublisher)
+        getphone = request.form.get('phone')
+        getid =request.form.get('id')
+        publisher = Publisher(name=getpublisher, phone_num=getphone, id=getid)
         db.session.add(publisher)
         flash(f'The Publisher {getpublisher} was added to your database', 'success')
         db.session.commit()
         return redirect(url_for('addpublisher'))
     return render_template('books/addpublisher.html', publisherss='publisherss')
 
+
 @app.route('/updatepublisher/<int:id>', methods=['GET', 'POST'])
 def updatepublisher(id):
     if 'email' not in session:
         flash('Please login first', 'danger')
         return render_template(url_for('login'))
-    updatepublisher = Publisher.query.get_or_404(id)
+    updatepublisher = db.session.query(Publisher).filter_by(id=id).first_or_404()
     publisher = request.form.get('publisher')
+    getphone = request.form.get('phone')
     if request.method == 'POST':
         updatepublisher.name = publisher
+        updatepublisher.phone_num = getphone
         flash('Your publisher has been updated', 'success')
         db.session.commit()
         return redirect(url_for('publishers'))
@@ -58,7 +65,7 @@ def deletepublisher(id):
     if 'email' not in session:
         flash('Please login first', 'danger')
         return render_template(url_for('login'))
-    publisher = Publisher.query.get_or_404(id)
+    publisher = db.session.query(Publisher).filter_by(id=id).first_or_404()
     if request.method == 'POST':
         db.session.delete(publisher)
         flash(f'The publisher {publisher.name} was deleted from your database', 'success')
@@ -70,8 +77,8 @@ def deletepublisher(id):
 @app.route('/category/<int:id>')
 def get_category(id):
     page = request.args.get('page', 1, type=int)
-    pbycategory = Book.query.filter_by(category_id=id).paginate(page=page, per_page=3)
-    return render_template('books/index.html', books=pbycategory, title=Category.query.get_or_404(id).name)
+    pbycategory = db.session.query(Book).filter_by(category_id=id).paginate(page=page, per_page=3)
+    return render_template('books/index.html', books=pbycategory, title=db.session.query(Category).get_or_404(id).name)
 
 @app.route('/addcategory', methods=['GET', 'POST'])
 def addcategory():
@@ -80,7 +87,8 @@ def addcategory():
         return redirect(url_for('login'))
     if request.method == 'POST':
         getcategory = request.form.get('category')
-        category = Category(name=getcategory)
+        getid = request.form.get('id')
+        category = Category(name=getcategory, id=id)
         db.session.add(category)
         flash(f'The Category {getcategory} was added to your database', 'success')
         db.session.commit()
@@ -92,7 +100,7 @@ def updatecategory(id):
     if 'email' not in session:
         flash('Please login first', 'danger')
         return render_template(url_for('login'))
-    updatecategory = Category.query.get_or_404(id)
+    updatecategory = db.session.query(Category).filter_by(id=id).first_or_404()
     category = request.form.get('category')
     if request.method == 'POST':
         updatecategory.name = category
@@ -108,7 +116,7 @@ def deletecategory(id):
         flash('Please login first', 'danger')
         return render_template(url_for('login'))
     
-    category = Category.query.get_or_404(id)
+    category = db.session.query(Category).filter_by(id=id).first_or_404()
     if request.method == 'POST':
         db.session.delete(category)
         flash(f'The category {category.name} was deleted from your database', 'success')
@@ -120,8 +128,8 @@ def deletecategory(id):
 @app.route('/author/<int:id>')
 def get_author(id):
     page = request.args.get('page', 1, type=int)
-    pbyauthor = Book.query.filter_by(author_id=id).paginate(page=page, per_page=3)
-    return render_template('books/index.html', books=pbyauthor, title=Author.query.get_or_404(id).name)
+    pbyauthor = db.session.query(Book).filter_by(author_id=id).paginate(page=page, per_page=3)
+    return render_template('books/index.html', books=pbyauthor, title=db.session.query(Author).get_or_404(id).name)
 
 @app.route('/addauthor', methods=['GET', 'POST'])
 def addauthor():
@@ -129,10 +137,14 @@ def addauthor():
         flash('Please login first', 'danger')
         return redirect(url_for('login'))
     if request.method == 'POST':
-        getauthor = request.form.get('author')
-        author = Author(name=getauthor)
+        getid = request.form.get('id')
+        getfn = request.form.get('fn')
+        getmn = ''
+        getln = request.form.get('ln')
+        getphone = request.form.get('phone')
+        author = Author(id=getid, Fname=getfn, Mname=getmn, Lname=getln, phone_num=getphone)
         db.session.add(author)
-        flash(f'The author {getauthor} was added to your database', 'success')
+        flash(f'The author {getfn} {getmn} {getln} was added to your database', 'success')
         db.session.commit()
         return redirect(url_for('addauthor'))
     return render_template('books/addpublisher.html', authorss='authorss')
@@ -142,10 +154,16 @@ def updateauthor(id):
     if 'email' not in session:
         flash('Please login first', 'danger')
         return render_template(url_for('login'))
-    updateauthor = Author.query.get_or_404(id)
-    author = request.form.get('author')
+    updateauthor = db.session.query(Author).filter_by(id=id).first_or_404()
+    getfn = request.form.get('fn')
+    getmn = ''
+    getln = request.form.get('ln')
+    getphone = request.form.get('phone')
     if request.method == 'POST':
-        updateauthor.name = author
+        updateauthor.Fname = getfn
+        updateauthor.Mname = getmn
+        updateauthor.Lname = getln
+        updateauthor.phone_num = getphone
         flash('Your author has been updated', 'success')
         db.session.commit()
         return redirect(url_for('authors'))
@@ -157,7 +175,7 @@ def deleteauthor(id):
     if 'email' not in session:
         flash('Please login first', 'danger')
         return render_template(url_for('login'))
-    author = Author.query.get_or_404(id)
+    author = db.session.query(Author).filter_by(id).first_or_404()
     if request.method == 'POST':
         db.session.delete(author)
         flash(f'The author {author.name} was deleted from your database', 'success')
@@ -166,9 +184,60 @@ def deleteauthor(id):
     flash(f'The author {author.name} can\'t be deleted', 'warning')
     return redirect(url_for('authors'))
 
+@app.route('/adddiscount', methods=['GET', 'POST'])
+def adddiscount():
+    if 'email' not in session:
+        flash('Please login first', 'danger')
+        return redirect(url_for('login'))
+    if request.method == 'POST':
+        getid = request.form.get('id')
+        getname = request.form.get('name')
+        getvalue = request.form.get('value')
+        discount = Discount(id=getid, name=getname, value=float(getvalue))
+        db.session.add(discount)
+        flash(f'The discount {getname} was added to your database', 'success')
+        db.session.commit()
+        return redirect(url_for('adddiscount'))
+    return render_template('books/addpublisher.html', discountss='discountss')
+
+@app.route('/updatediscount/<int:id>', methods=['GET', 'POST'])
+def updatediscount(id):
+    if 'email' not in session:
+        flash('Please login first', 'danger')
+        return render_template(url_for('login'))
+    updatediscount = db.session.query(Discount).filter_by(id=id).first_or_404()
+    getfn = request.form.get('fn')
+    getmn = ''
+    getln = request.form.get('ln')
+    getphone = request.form.get('phone')
+    if request.method == 'POST':
+        updatediscount.Fname = getfn
+        updatediscount.Mname = getmn
+        updatediscount.Lname = getln
+        updatediscount.phone_num = getphone
+        flash('Your discount has been updated', 'success')
+        db.session.commit()
+        return redirect(url_for('discounts'))
+        
+    return render_template('books/updatepublisher.html', title='Update discount page', updatediscount=updatediscount)
+
+@app.route('/deletediscount/<int:id>', methods=['POST'])
+def deletediscount(id):
+    if 'email' not in session:
+        flash('Please login first', 'danger')
+        return render_template(url_for('login'))
+    discount = db.session.query(Discount).filter_by(id=id).first_or_404()
+    if request.method == 'POST':
+        db.session.delete(discount)
+        flash(f'The discount {discount.name} was deleted from your database', 'success')
+        db.session.commit()
+        return redirect(url_for('discounts'))
+    flash(f'The discount {discount.name} can\'t be deleted', 'warning')
+    return redirect(url_for('discounts'))
+
 @app.route('/book/<int:id>')
 def single_book(id):
-    book = Book.query.get_or_404(id)
+    book = db.session.query(Book).get_or_404(id)
     return render_template('books/single_book.html', book=book)
 
 @app.route('/addbook', methods=['POST', 'GET'])
@@ -176,50 +245,78 @@ def addbook():
     if 'email' not in session:
         flash('Please login first', 'danger')
         return redirect(url_for('login'))
-    publishers = Publisher.query.all()
-    categories = Category.query.all()
-    authors = Author.query.all()
+    publishers = db.session.query(Publisher).all()
+    categories = db.session.query(Category).all()
+    authors = db.session.query(Author).all()
+    discounts = db.session.query(Discount).all()
     form = Addbook(request.form)
     if request.method == 'POST':
-        name = form.name.data
+        isbn = form.isbn.data
+        title = form.title.data
+        edition = form.edition.data
+        publish_date = form.publish_date.data
         price = form.price.data
-        discount = form.discount.data
         stock = form.stock.data
-        desc = form.description.data
-        author = request.form.get('author')
+        sales_amount = form.sales_amount.data
+        shortdesc = form.shortdesc.data
+        longdesc = form.longdesc.data
+        author = request.form.getlist('author')
         publisher = request.form.get('publisher')
-        category = request.form.get('category')
+        category = request.form.getlist('category')
+        discount = request.form.getlist('discount')
         image_1 = photos.save(request.files.get('image_1'), name=secrets.token_hex(10) + '.')
         image_2 = photos.save(request.files.get('image_2'), name=secrets.token_hex(10) + '.')
         image_3 = photos.save(request.files.get('image_3'), name=secrets.token_hex(10) + '.')
         
-        addpro = Book(name=name, price=price, discount=discount, stock=stock, author_id=author, desc=desc, publisher_id=publisher, category_id=category, image_1=image_1, image_2=image_2, image_3=image_3)
+        addpro = Book(title=title, price=price, stock=stock, shortdesc=shortdesc,
+                      Pname=publisher, image_1=image_1, image_2=image_2, image_3=image_3, edition=edition,
+                      ISBN=isbn, publish_date=publish_date, sales_amount=sales_amount, longdesc=longdesc)
+
         db.session.add(addpro)
-        flash(f'The book {name} has been added to your database', 'success')
+        
+        for phone in author:
+            add_auth_book = Author_write_book(Aphone_num=phone, ISBN=isbn)
+            db.session.add(add_auth_book)
+            
+        for name in category:
+            add_cate_book = Book_belongs_to_category(category_name=name, ISBN=isbn)
+            db.session.add(add_cate_book)
+            
+        for id in discount:
+            add_book_discount = Book_discount(discount_id=int(id), book_id=isbn)
+            db.session.add(add_book_discount)
+        
+        flash(f'The book {title} has been added to your database', 'success')
         db.session.commit()
         return redirect(url_for('admin'))
-    return render_template('books/addbook.html', title='Add Book page', form=form, publishers=publishers, categories=categories, authors=authors)
+    return render_template('books/addbook.html', title='Add Book page', form=form, publishers=publishers, categories=categories, authors=authors, discounts=discounts)
 
-@app.route('/updatebook/<int:id>', methods=['GET', 'POST'])
-def updatebook(id):
-    publishers = Publisher.query.all()
-    categories = Category.query.all()
-    authors = Author.query.all()
-    book = Book.query.get_or_404(id)
+@app.route('/updatebook/<string:isbn>', methods=['GET', 'POST'])
+def updatebook(isbn):
+    if 'email' not in session:
+        flash('Please login first', 'danger')
+        return redirect(url_for('login'))
+    publishers = db.session.query(Publisher).all()
+    categories = db.session.query(Category).all()
+    authors = db.session.query(Author).all()
+    discounts = db.session.query(Discount).all()
+    book = db.session.query(Book).filter_by(ISBN=isbn).first_or_404()
     form = Addbook(request.form)
-    publisher = request.form.get('publisher')
-    category = request.form.get('category')
-    author = request.form.get('author')
+    # author = request.form.getlist('author')
+    # publisher = request.form.get('publisher')
+    # category = request.form.getlist('category')
+    # discount = request.form.getlist('discount')
     
     if request.method == 'POST':
-        book.name = form.name.data
+        book.ISBN = form.isbn.data
+        book.title = form.title.data
+        book.edition = form.edition.data
+        book.publish_date = form.publish_date.data
         book.price = form.price.data
-        book.discount = form.discount.data
         book.stock = form.stock.data
-        book.publisher_id = publisher
-        book.category_id = category
-        book.author_id = author
-        book.desc = form.description.data
+        book.sales_amount = form.sales_amount.data
+        book.shortdesc = form.shortdesc.data
+        book.longdesc = form.longdesc.data
         if request.files.get('image_1'):
             try:
                 os.unlink(os.path.join(current_app.root_path, 'static/images/book/' + book.image_1))
@@ -246,20 +343,24 @@ def updatebook(id):
         return redirect(url_for('admin'))
         
     
-    form.name.data = book.name
+    form.isbn.data = book.ISBN 
+    form.title.data = book.title
+    form.edition.data = book.edition
+    form.publish_date.data = book.publish_date 
     form.price.data = book.price
-    form.discount.data = book.discount
     form.stock.data = book.stock
-    form.description.data = book.desc
+    form.sales_amount.data = book.sales_amount
+    form.shortdesc.data = book.shortdesc
+    form.longdesc.data = book.longdesc
     return render_template('books/updatebook.html', title='Update book page', 
-                           form=form, publishers=publishers, categories=categories, authors=authors,book=book)
+                           form=form, publishers=publishers, categories=categories, authors=authors,book=book, discounts=discounts)
     
-@app.route('/deletebook/<int:id>', methods=['POST'])
-def deletebook(id):
+@app.route('/deletebook/<string:isbn>', methods=['POST'])
+def deletebook(isbn):
     if 'email' not in session:
         flash('Please login first', 'danger')
         return redirect(url_for('login'))
-    book = Book.query.get_or_404(id)
+    book = db.session.query(Book).filter_by(ISBN=isbn).first_or_404()
     if request.method == 'POST':
         try:
             os.unlink(os.path.join(current_app.root_path, 'static/images/book/' + book.image_1))
@@ -268,8 +369,8 @@ def deletebook(id):
         except Exception as e:
             print(e)
         db.session.delete(book)
-        flash(f'The book {book.name} was deleted from your database', 'success')
+        flash(f'The book {book.title} was deleted from your database', 'success')
         db.session.commit()
         return redirect(url_for('admin'))
-    flash(f'The book {book.name} can\'t be deleted', 'warning')
+    flash(f'The book {book.title} can\'t be deleted', 'warning')
     return redirect(url_for('admin'))
